refactor(omo-senpi): group Kibitzer modules by ownership
