+++ packages/omo-senpi/src/components/memory/kibitzer-runner-completion.test.ts
@@ -2,7 +2,7 @@ import { afterEach, describe, expect, test } from "bun:test"
 import { mkdir, readdir, readFile, writeFile } from "node:fs/promises"
 import { join } from "node:path"
 import type { ChildSpec, InProcessRunnerLike } from "@oh-my-opencode/senpi-task"
-import { KIBITZER_NUDGE_TOOL_NAME, type KibitzerNudgeTool } from "./kibitzer-nudge-tool"
+import { KIBITZER_NUDGE_TOOL_NAME, type KibitzerNudgeTool } from "./kibitzer/nudge-tool"
 import { KibitzerGateRunner } from "./kibitzer-runner"
 import { CANDIDATE_PATH, callNudge, fixture, launchInput, nudgeOnce, roots, runnerOptions, scriptedSession } from "./kibitzer-runner.test-support"
 import { rmEfaultTolerant } from "./teardown.test-support"
