+++ packages/omo-senpi/src/components/memory/kibitzer-runner.ts
@@ -37,7 +37,7 @@ import type { ComponentLogger } from "../../extension/types"
 import type { SenpiOmoConfigResult } from "../config-resolution"
 import type { RecallTranscriptTurn } from "./recall-wiring"
 import { resolveReflectionModel } from "./worker/resolve-model"
-import { normalizeGateReason } from "./kibitzer-judge-outcome"
+import { normalizeGateReason } from "./kibitzer/judge-outcome"
 import { runKibitzerJudge } from "./kibitzer-judge-run"
 import { abortAndDispose } from "./kibitzer-lifecycle"
 import { writeKibitzerRunOutcome } from "./kibitzer-run-retention"
