+++ packages/omo-senpi/src/components/memory/kibitzer-trigger.test.ts
@@ -3,8 +3,8 @@ import { rm } from "node:fs/promises"
 
 import { buildIdentityPaths, PendingNudges, RecallLedger } from "@oh-my-opencode/memory-core"
 import { FakeExtensionAPI } from "../../../test-support/fake-extension-api"
-import { createKibitzerDelivery } from "./kibitzer-delivery"
-import { registerKibitzerHooks } from "./kibitzer-hooks"
+import { createKibitzerDelivery } from "./kibitzer/delivery"
+import { registerKibitzerHooks } from "./kibitzer/hooks"
 import { memorySettings } from "./memory.test-support"
 import { createMemoryRecallWiring } from "./recall-wiring"
 import { beforeAgentStart, eventContext, fixture, KUBERNETES_PROMPT, ROLLOUTS_PATH, SESSION_ID, userEntry, type BranchEntry } from "./recall-wiring.test-support"
