+++ packages/omo-senpi/src/components/memory/kibitzer/composition.ts
@@ -1,18 +1,18 @@
 import { PendingNudges, RecallLedger } from "@oh-my-opencode/memory-core"
 
-import type { ComponentContext, ComponentLogger, SenpiExtensionAPI } from "../../extension/types"
-import type { MemoryIdentityContext } from "./context"
-import { createKibitzerDelivery, type KibitzerDelivery } from "./kibitzer-delivery"
-import { registerKibitzerHooks } from "./kibitzer-hooks"
-import { createThrottledPrune } from "./kibitzer-run-retention"
-import { createKibitzerTrigger, type KibitzerTrigger } from "./kibitzer-trigger"
-import { createKibitzerGateWiring, type KibitzerGateWiring } from "./kibitzer-wiring"
-import { sessionIdFrom } from "./wiring-context"
-import { resolveMemoryModelRegistry } from "./model-registry-resolver"
-import { ToolArgWindow } from "./recall-query-planner-tools"
-import type { MemoryRecallWiring } from "./recall-wiring"
-import type { MemoryRuntimeWiring } from "./wiring-runtime"
-import type { MemoryWiringOptions } from "./wiring-types"
+import type { ComponentContext, ComponentLogger, SenpiExtensionAPI } from "../../../extension/types"
+import type { MemoryIdentityContext } from "../context"
+import { createKibitzerDelivery, type KibitzerDelivery } from "./delivery"
+import { registerKibitzerHooks } from "./hooks"
+import { createThrottledPrune } from "../kibitzer-run-retention"
+import { createKibitzerTrigger, type KibitzerTrigger } from "../kibitzer-trigger"
+import { createKibitzerGateWiring, type KibitzerGateWiring } from "../kibitzer-wiring"
+import { sessionIdFrom } from "../wiring-context"
+import { resolveMemoryModelRegistry } from "../model-registry-resolver"
+import { ToolArgWindow } from "../recall-query-planner-tools"
+import type { MemoryRecallWiring } from "../recall-wiring"
+import type { MemoryRuntimeWiring } from "../wiring-runtime"
+import type { MemoryWiringOptions } from "../wiring-types"
 
 export interface KibitzerComposition {
   readonly gate: KibitzerGateWiring
