+++ packages/omo-senpi/src/components/memory/kibitzer/delivery-idle.test.ts
@@ -5,15 +5,15 @@ import { join } from "node:path"
 
 import { buildIdentityPaths, PendingNudges, RecallLedger, type RecallNudge } from "@oh-my-opencode/memory-core"
 
-import { IdleInjectionCoordinator } from "../../extension/idle-injection-coordinator"
-import { createMemoryIdentityContext } from "./context"
-import { createMemoryBinding } from "./binding"
-import { createKibitzerDelivery, type KibitzerDelivery } from "./kibitzer-delivery"
-import { FakeExtensionAPI } from "../../../test-support/fake-extension-api"
-import { registerKibitzerHooks } from "./kibitzer-hooks"
-import { createKibitzerTrigger } from "./kibitzer-trigger"
-import { ToolArgWindow } from "./recall-query-planner-tools"
-import { beforeAgentStart, eventContext } from "./recall-wiring.test-support"
+import { IdleInjectionCoordinator } from "../../../extension/idle-injection-coordinator"
+import { createMemoryIdentityContext } from "../context"
+import { createMemoryBinding } from "../binding"
+import { createKibitzerDelivery, type KibitzerDelivery } from "./delivery"
+import { FakeExtensionAPI } from "../../../../test-support/fake-extension-api"
+import { registerKibitzerHooks } from "./hooks"
+import { createKibitzerTrigger } from "../kibitzer-trigger"
+import { ToolArgWindow } from "../recall-query-planner-tools"
+import { beforeAgentStart, eventContext } from "../recall-wiring.test-support"
 
 const SESSION_ID = "idle-session"
 const NUDGE: RecallNudge = { path: "reference/idle.md", hint: "Use the idle wake path." }
