+++ packages/omo-senpi/src/components/memory/kibitzer/observability.test.ts
@@ -4,18 +4,18 @@ import { join } from "node:path"
 import { tmpdir } from "node:os"
 
 import { buildIdentityPaths, PendingNudges, RecallLedger, type RecallNudge } from "@oh-my-opencode/memory-core"
-import { IdleInjectionCoordinator } from "../../extension/idle-injection-coordinator"
-import { createMemoryBinding } from "./binding"
-import { createMemoryIdentityContext, type MemoryIdentityContext } from "./context"
-import { tryAcquireJudgeSlot } from "./kibitzer-concurrency"
-import { createKibitzerDelivery } from "./kibitzer-delivery"
-import { KibitzerGateRunner } from "./kibitzer-runner"
-import { CANDIDATES, nudgeOnce, registrySnapshot, runnerOptions, scriptedSession } from "./kibitzer-runner.test-support"
-import { createKibitzerTrigger } from "./kibitzer-trigger"
-import { createKibitzerGateWiring } from "./kibitzer-wiring"
-import { MemoryFakeExtensionAPI, memorySettings } from "./memory.test-support"
-import { ToolArgWindow } from "./recall-query-planner-tools"
-import { createRecallDrain } from "./recall-drain"
+import { IdleInjectionCoordinator } from "../../../extension/idle-injection-coordinator"
+import { createMemoryBinding } from "../binding"
+import { createMemoryIdentityContext, type MemoryIdentityContext } from "../context"
+import { tryAcquireJudgeSlot } from "../kibitzer-concurrency"
+import { createKibitzerDelivery } from "./delivery"
+import { KibitzerGateRunner } from "../kibitzer-runner"
+import { CANDIDATES, nudgeOnce, registrySnapshot, runnerOptions, scriptedSession } from "../kibitzer-runner.test-support"
+import { createKibitzerTrigger } from "../kibitzer-trigger"
+import { createKibitzerGateWiring } from "../kibitzer-wiring"
+import { MemoryFakeExtensionAPI, memorySettings } from "../memory.test-support"
+import { ToolArgWindow } from "../recall-query-planner-tools"
+import { createRecallDrain } from "../recall-drain"
 
 const sessionId = "observability-session"
 const nudge: RecallNudge = { path: "memory/rollouts.md", hint: "Drain nodes before rollout." }
