+++ packages/omo-senpi/src/components/memory/wiring-static.ts
@@ -19,7 +19,7 @@ import { registerSkillsUsage, type SkillsUsageTracker } from "./skills-usage"
 import { registerMemoryUsage, type MemoryUsageTracker } from "./memory-usage"
 import type { createMemoryNoticeWiring } from "./memory-notice-wiring"
 import type { KibitzerGateWiring } from "./kibitzer-wiring"
-import type { KibitzerComposition } from "./wiring-kibitzer"
+import type { KibitzerComposition } from "./kibitzer/composition"
 import type { createMemoryRecallWiring } from "./recall-wiring"
 import { createReflectionTriggerWiring } from "./trigger-wiring"
 import { registerMemoryToolSurface } from "./tools"
